Move refineData into the schema code (to avoid cycles)
